requirements.txt
langchain-core>=0.1.0
langchain-experimental>=0.0.47
langchain-openai>=0.0.5
langgraph>=0.0.15
matplotlib>=3.7.0
langchain-community>=0.0.11
python-dotenv>=1.0.0
e2b-code-interpreter>=0.10.2
pytz>=2023.3